requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langgraph>=0.0.20
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
fake_useragent>=1.2.0
playwright>=1.40.0
tiktoken>=0.5.0

# Search engines
googlesearch-python>=1.2.3
wikipedia>=1.4.0
arxiv>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Document parsing
pypdf2>=3.0.0
python-docx>=1.0.0

# Vector store and embeddings
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0

# Testing
pytest>=7.0.0
